MM-21143: implement SimpleController.reload
Added all the methods that we currently have that happens
when a user performs a reload.

Marked the rest of them as TODO.

While at it, we also move the getter methods from MutableUserStore
to UserStore to be able to access from the controller.

Tested with this piece of code. (Needs the server running)

func TestReload(t *testing.T) {
	c := SimpleController{}
	c.Init(userentity.New(memstore.New(), 0, userentity.Config{
		ServerURL:    "http://localhost:8065",
		WebSocketURL: "ws://localhost:8065",
	}))

	status := c.signUp()
	t.Logf("%+v", status)

	status = c.login()
	t.Logf("%+v", status)

	status = c.reload()
	t.Logf("%+v", status)
}