MM-21367: Fix panic when Disconnect gets called consecutively (#74)
* MM-21367: Fix panic when Disconnect gets called consecutively

The issue was that when the controller loop runs for the 2nd time,
the user does not login again, and the API throws "user already signed up error".
This means user.Connect does not get called. Now when it goes forward,
it calls Disconnect again during logout, essentially calling disconnect 2 times
consecutively.

This creates a panic because we call `close(ue.wsClosing)` twice consecutively.

We fix 2 bugs here:

1. We reset the channels in Connect.
2. We add a state check to indicate whether the user is disconnected/connected.
And depending on that, bail out early from the disconnect function.

* Address review comments

- Now that we are using a boolean flag to check for connectedness,
we can get rid of the mutex to control access to the websocket connection.
And close the connection inside the for-select loop itself.

* Remove WebsocketClient from UserEntity

Now the entire connection logic is handled in a single goroutine.
Hence there is no need to make it part of the struct.