MM-21505: implement LoadTester.Status
We add various counters and state variables to the struct
and expose them through a method.

This will help us get more information about the running load test
and store the data to be analyzed later.

While at it, we also add a AddUser/RemoveUser call in main.go
to help test race conditions. I had to do this every time manually
and thought it would be good to just keep it there.