Fix race in reading statusChan
While stopping a LoadTester, we just stopped reading from the channel
without ensuring the goroutine exited. This cause a race condition
when re-starting an existing loadtester.

We fix this by sending a quit signal which would actually exit
from the goroutine and ensure the channel read is no longer in memory.