MM-23183: fix deadlock in emulateUserTyping
In rare cases, the random duration in time.Ticker be zero, and
then it would get stuck because it returns a nil channel if d <=0.
This would prevent the controller from moving forward in it's loop,
and eventually not even receive the stop signal. We fix this
by simply changing it to a time.Sleep. This is much better and also
bypasses the problem of the ticker not being garbage collected.

Fun debugging session. Moral of the story: don't try complicated things
when simple works :)

We also fix the missing case of not calling the cleanup function.