MM-23262: Use the cluster name as a prefix for keypair
This gets rid of the hardcoding. A general problem still persists
where if two people use the same cluster name, then there will be a clash again.

But then the problem with adding a random string is that we need to pass the exact
same thing during destruction too.

So for now, we will recommend users to prefix the cluster names with their names
to compartmentalize deployments.