MM-23127: Fix JSON parse error with TF_LOG set
When TF_LOG would be set, there would be another
stray log line which would be part of the JSON output for
"terraform output -json". This would fail the JSON parsing.

We fix this by removing anything after the last }. This is a hack
and we have to resort to this because we are forced to parse
output from the CLI rather than go through an API.

This will inevitably break if the TF_LOG output contains a } too.
Then we can resort to even nastier hacks.