MM-23126: Validate that keys are added to the ssh agent
The ssh agent creation would succeed even if there are no keys
added to the agent. This would lead to the deployer eventually getting stuck.

We add an additional check to list the keys first and verify that at least one
key is there.

This required the connection creation to be split up by first creating an agent
and then using that to create a connection.