MM-23422: Fix order of apt-get update
I couldn't fully figure out the cause of this. This seems
to be related to terraform implementation of how it runs
inline commands.

The best case scenario is to add the sanest hacks. According
to my tests, this makes everything go smooth. Although
there is a chance of failing again because the root cause
of failure is still undetected.

We first run apt-get update before running any command.
Then we run it again after adding the key to the keyring.
Somehow apt-get update fails to update the list completely after
the key gets added.