MM-24910: Rewrite ws client
This is a complete rewrite of the websocket client stripping out the parts
that aren't necessary to make it a extremely simple, bare-bones and safe
implementation.

There is just one listener goroutine and nothing else. No timeouts, no ping handlers,
no response channels. No extra frills.

Writes are synchronized through a mutex and close is made synchronous by waiting
on the reader waitgroup to be done.