MM-25511: Improve nginx keepalive connections (#297)
1024 is too high a number for it. I misunderstood the directive to be something
where nginx would only keep connections alive when the client wouid be done with them.
But this is a different set of connections than "active" ones. And it is specifically mentioned in the docs.

> It should be particularly noted that the keepalive directive does not limit the total number
of connections to upstream servers that an nginx worker process can open.
The connections parameter should be set to a number small enough to let upstream servers
process new incoming connections as well.

https://nginx.org/en/docs/http/ngx_http_upstream_module.html#keepalive

After some trial and error, I found 64 to be a decent number.

Previously timeout errors would start to appear at 1k connections, now it comes at around 5k.