MM-25437: Shutdown agents in parallel
While shutting down a huge cluster of more than 10k users,
the stopping of the coordinator service will time out, leading
to the service being forcibly killed but the agents would still
be running.

We shutdown in parallel to reduce the time.

The DefaultTimeoutStopSec is 90 seconds, so we can keep that as is.