MM-26416: move Prometheus queries to load from config
This PR changes the graph queries from being hardcoded
to being loadable from config. This takes care of using custom labels
for the queries and at the same time, use any custom query that the user
wants to run.

We also add a Label field for the Avg and P99 queries.

While here, we fix an unrelated bug that did not close the file.

Docs are updated and sample config file is also changed.