Set postgres:default_text_search_config to english
Aurora PostgreSQL sets the default_text_search_config parameter to
'simple', but 'english' is assumed in many default places of the
codebase; e.g., the GIN indexes have a hard-coded 'english' value, so
text search queries should use english if possible.

This commit sets this default parameter to 'english' in the
creation phase of the deployment, only if the instance engine is
aurora-postgresql.