Use -chdir instead of passing a dir to terraform
Passing a dir as an argument of a terraform command was deprecated in
0.14 and removed in 0.15. We need to use the global option -chdir
instead, which, as all global options, needs to appear before the
subcommand.

See
https://www.terraform.io/cli/v1.1.x/commands/init#passing-a-different-configuration-directory
for more information.

This change forces us to store the directory of the state file as well,
since the working directory is no longer the same as the directory we're
running the command in (which is where the state file lives).