MM-47262: Bump to v1.3.3 (#516)
* Bump to v1.3.3

Also, change the notion of supported version to minimum supported
version, which is defined as the latest version tested. According to the
v1 compatibility promises[1], later v1.x versions will be compatible, so
we should not limit usage of later versions. When v2 is released, we can
revisit this check.

[1]: https://www.terraform.io/language/v1-compatibility-promises

* Check for major versions greater than 1

Rename minimumRequiredVersion to requiredVersion and document what we
actually mean by that

* Specify new version requirements in the docs