Base all Terraform commands on the assets dir
Before this commit, all Terraform commands were based on a temporary
directory to which all Terraform source files were copied.
This worked well when using Terraform v0.14, since the .terraform
directory was still being created in the root of the repository (where
the go run ./cmd/ltctl deployment create command was run).
When bumping to v0.15, we started using the global option -chdir
instead of the deprecated way of passing the directory as the last
argument to the command. This procedure still used a temporary
directory, but it had an unexpected side effect: the .terraform
directory was also created under that temporary directory, instead of
under the directory from where the command was being run.

This commit revamps this whole setup. We're still using the global
-chdir option, but it's no longer using a temporary directory, but
hardcoded to deployment/terraform/assets, where all the Terraform source
files already live. This is what Terraform expects in a default
environment: you run the command where your .tf files live.