Add new config to hold a Terraform state directory
Instead of relying on the repository file hierarchy, use a
user-configured directory as the working directory for Terraform, to
mange the source files, the .terraform directory and state files.

This defaults to /var/lib/mattermost-load-test-ng. /var/lib, as per the
Filesystem Hierarchy Standard, "holds state information pertaining to an
application or the system. [...]State information is generally used to
preserve the condition of an application (or a group of inter-related
applications) between invocations and between different instances of the
same application." This makes it the perfect candidate for storing the
state of Terraform, which we need between invocations of different
deployment subcommands.

This directory is created if it does not exist when calling
terraform.New(), so that function can now fail and its error needs to be
managed by all callers.