MM-47965: Revamp working directory management (#517)
* Base all Terraform commands on the assets dir

Before this commit, all Terraform commands were based on a temporary
directory to which all Terraform source files were copied.
This worked well when using Terraform v0.14, since the .terraform
directory was still being created in the root of the repository (where
the go run ./cmd/ltctl deployment create command was run).
When bumping to v0.15, we started using the global option -chdir
instead of the deprecated way of passing the directory as the last
argument to the command. This procedure still used a temporary
directory, but it had an unexpected side effect: the .terraform
directory was also created under that temporary directory, instead of
under the directory from where the command was being run.

This commit revamps this whole setup. We're still using the global
-chdir option, but it's no longer using a temporary directory, but
hardcoded to deployment/terraform/assets, where all the Terraform source
files already live. This is what Terraform expects in a default
environment: you run the command where your .tf files live.

* Move .terraform.lock.hcl to the assets directory

As everything's now based on the deployment/terraform/assets directory,
it makes sense to move the lock file there as well.

* Remove the generated assets

Assets can be retrieved from the directory itself, so we don't need to
use the gobindata stuff anymore for this.

* Remove the Cleanup function, no longer needed

The only thing added to the Terraform working directory (now
deployment/terraform/assets) is the .terraform directory, which can live
there from one run to another. If the lock file changes, Terraform will
take care of updating what's needed inside that directory.

* Remove stateDir value

All Terraform commands are now based on the deployment/terraform/assets
directory, so instead of forcing the terraform.tfstate (and backup) file
to live in the repository root (where the go run ./cmd/ltctl deployment
command is run), we can let it live in the same base directory for
everything else.

* Add new config to hold a Terraform state directory

Instead of relying on the repository file hierarchy, use a
user-configured directory as the working directory for Terraform, to
mange the source files, the .terraform directory and state files.

This defaults to /var/lib/mattermost-load-test-ng. /var/lib, as per the
Filesystem Hierarchy Standard, "holds state information pertaining to an
application or the system. [...]State information is generally used to
preserve the condition of an application (or a group of inter-related
applications) between invocations and between different instances of the
same application." This makes it the perfect candidate for storing the
state of Terraform, which we need between invocations of different
deployment subcommands.

This directory is created if it does not exist when calling
terraform.New(), so that function can now fail and its error needs to be
managed by all callers.

* Revert "Remove the generated assets"

This reverts commit 2475e28bd5633d64c3cc08dbef98da8f007c9f6b.

* Make assets

* Document use of the new setting, improve error msg

* Make linter happy