Add S3BucketDumpURI setting to comparison command
Similar to the setting in the deployment configuration, this commit adds
an equally named setting to the comparison configuration, allowing
comparison tests to pre-populate the deployed bucket with data from a
publicly accessible bucket.

As before, the copy happens locally, invoking `aws s3 cp` in the machine
running the ltctl binary. Before copying the data, the code in the
comparison command makes sure to delete the contents of the destination
bucket first, to allow for repeatable tests.