Add S3BucketDumpURI setting to deployer config
This setting, when non-empty, will be used to pre-populate the bucket
created during the deployment (which happens only if there's more than
one app instance). This unblocks load-tests to use not only a DB dump,
but the corresponding S3 bucket associated to it, so that it can
recreate any arbitrary state previously stored.

There is no way in Terraform to copy the contents of one bucket to
another (there's `aws_s3_object_copy`, but that copies object by object
reading them into memory, so it's a waste of time and memory), that's
why we defer to use a null_resource that simply executes the `aws s3 cp`
command locally. This will add a new requirement for the user: they need
to have the AWS CLI installed, or the command will fail.