MM-46435: Add configuration settings to prepopulate deployed S3 bucket (#524)
* Add s3uri validation for config settings

With this addition, config structs can use the validate:"s3uri" tag in
their fields to check that the value used is indeed an S3 URI, i.e., a
URI accepted by url.ParseRequestURI with `s3` as its scheme.

* Add S3BucketDumpURI setting to deployer config

This setting, when non-empty, will be used to pre-populate the bucket
created during the deployment (which happens only if there's more than
one app instance). This unblocks load-tests to use not only a DB dump,
but the corresponding S3 bucket associated to it, so that it can
recreate any arbitrary state previously stored.

There is no way in Terraform to copy the contents of one bucket to
another (there's `aws_s3_object_copy`, but that copies object by object
reading them into memory, so it's a waste of time and memory), that's
why we defer to use a null_resource that simply executes the `aws s3 cp`
command locally. This will add a new requirement for the user: they need
to have the AWS CLI installed, or the command will fail.

* Add AWS profile as a variable in deployment

Until now, the `mm-loadtest` profile was hardcoded in cluster.tf, and
that string was not easily recoverable from the code. Converting it into
a variable achievees two goals:
- It lets users specify a different profile name if required.
- The code can now access the value in the config if needed.

* Add S3BucketDumpURI setting to comparison command

Similar to the setting in the deployment configuration, this commit adds
an equally named setting to the comparison configuration, allowing
comparison tests to pre-populate the deployed bucket with data from a
publicly accessible bucket.

As before, the copy happens locally, invoking `aws s3 cp` in the machine
running the ltctl binary. Before copying the data, the code in the
comparison command makes sure to delete the contents of the destination
bucket first, to allow for repeatable tests.

* Use CommandContext to manage early returns

* Buffer the channel so that the routine can return