Avoid a nil pointer dereference if RoundTrip fails
There are two alternatives to solve this:
- What this commit does: use 0 as statusCode when RoundTrip fails. This
gives more information to the metric, since 0 is not a valid HTTP status
code, but if the users don't know this, seeing a 0 there can be confusing.
- Observe the metric only if t.transport.RoundTrip(req) finishes
successfully. This avoids the confusion with a 0 statusCode, but it also
skips all failing round trips, which most probably fail because of a
timeout, which are useful to see in the metric.