MM-53560: Do not set param group name if not defined
We were unconditionally setting the param group name
for an RDS cluster even when there were no parameters defined.

This would lead to an issue re-creating the cluster because
the param group woudn't really exist, but the db cluster
got created. Ideally, this should have thrown an error from
Terraform while trying to attach a param group that doesn't
exist. But apparently it can indeed attach a dummy group.

So we are left with manually checking if we need to set
a group or not.

https://mattermost.atlassian.net/browse/MM-53560