MM-53845: Attach to an existing rds cluster (#638)
This is not completely straightforward and there
might be better ways to achieve this, but this feels
like a reasonable first step.

Our first problem is to be able to import the restored
cluster into our terraform state and add instances
to the cluster.

My initial approach was to add the import block into the
terraform file itself, but the main challenge was that
the import block does not allow targets to be variables.

So I had to move the command to be run as a CLI.

The next challenge was to attach security groups
once they are imported. This was necessary because
if a cluster is imported, rather than created, then
the security groups don't get attached by Terraform
automatically since the block is entirely skipped.

So we have to spawn a shell command to attach the security
groups.

Once these are done, terraform works as expected and
idempotently. I have tested with deploying multiple times
and it works fine.

The only catch is that destroying a setup does not
delete the cluster. That one we have to manually.
Don't think that should be a big deal. Again, we can
write another shell command to delete it, but leaving it
out for now.

https://mattermost.atlassian.net/browse/MM-53845