Stop/Start DB Cluster after load test
We add the ability to stop/start the DB cluster after a load-test
is complete.

There is a dedicated command that the users can run any time.
Additionally, there is also an automated way to stop the DB
after a load test finishes, which is exposed via a flag.

Using the flag makes the load test runs synchronously, so it is
advised to be used only by advanced users running the command
preferably inside an EC2 instance in a screen session.

https://mattermost.atlassian.net/browse/MM-52240