Remove Cloudwatch log policy
There's a limit of 10 Cloudwatch log policies per region per account, so
it's not scalable to create one with every deployment. Instead, we rely
on such a policy already present in the AWS account. If it is not
present, the only downside is that logs cannot be viewable through
Cloudwatch, but everything else should keep working.