Revert the disabling of proxy_cache_lock
Release testing for v1.19 of the load-test tool showed that disabling
proxy_cache_lock was affecting the performance of the app nodes, which
were eventually killed by the overhead of the requests that were no
longer waiting for the cache to be updated.

This commit reverts that change, and a subsequent investigation into the
proxy_cache_lock_timeout fine-tuning will follow.
See https://mattermost.atlassian.net/browse/MM-59871