MM-60548: Use DB instance endpoints directly
AWS enforces a limit of 5 custom endpoints per cluster:
https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/Aurora.Endpoints.Custom.html

I am not sure how we didn't run into this in our
previous ceiling tests, but anyways I got blocked
by this randomly.

Turns out that we don't need custom endpoints in the
first place, because we were anyways mapping 1:1 from
endpoint to instance. So we can directly use the
instance endpoint.

This also allows us to use the writer field in the
instance struct to let us cleanly identify the writer
rather than looking for an instance prefix.

https://mattermost.atlassian.net/browse/MM-60548