MM-52099: Be more gentle with comparison failures (#820)
* Remove unneeded, very old code

This TODO was fixed in
https://github.com/mattermost/mattermost-load-test-ng/pull/395
four years ago now.

* Remove the cancel channel

This was used as a communication channel between completely independent
runs, which don't need to be canceled if other runs fail. So we can
simply get rid of its complexity.

* Note individual deployment failures in the results

Before, we didn't generate results at all if one test failed. Now, we
need to explicitly tell the user that although other tests succeeded,
some of them failed.

* Specify the deployment ID in error logs