Remove import of the restored RDS cluster
It turns out this import has been silently failing forever. Starting
with Terraform v1.9, though, it started triggering an error that made
the deployment fail. For more information, see
https://developer.hashicorp.com/terraform/language/upgrade-guides#invalid-import-blocks

But the thing is that we don't actually need it! We already create the
RDS cluster conditionally, skipping it if ClusterIdentifier is
specified. Also, the DB instances are attached to the new cluster if one
is created but, again, if ClusterIdentifier is specified, they are
attached to that one.

That's why the whole thing has always worked despite the import command
failing: we don't actually need it.