Try another service env when validating licenses
A common error when providing a license is to use a different service
environment (dev, test or production) to the one the license was
generated for. This error is not surfaced by the call to
ValidateLicense, so we explicitly test a different environment to the
one currently in use to check whether license would be valid there. If
that's the case, then we return a much more helpful error to the user.