MM-60036: License validation (#832)
* Pass []byte to validateLicense instead of a path

This is to make testing easier. We can read the file outside of the
function, no big deal.

* Try another service env when validating licenses

A common error when providing a license is to use a different service
environment (dev, test or production) to the one the license was
generated for. This error is not surfaced by the call to
ValidateLicense, so we explicitly test a different environment to the
one currently in use to check whether license would be valid there. If
that's the case, then we return a much more helpful error to the user.

* Test ValidateLicense

Double-check that the error returned in case of a mismatch in the
service environment is the one we expect.

* Revert "Pass []byte to validateLicense instead of a path"

This reverts commit a7aac39fa29005889251a201b5174048f6fbc7ee.

* Use an asset file in TestValidateLicense