Use the correct type in reflect.MakeSlice
MakeSlice expects its first argument to be the type of a slice, not the
type of the underlying elements of a slice; i.e., we need to pass
`[]string`, not `string`.

Before this commit, we were passing the type of the slice's Elem(),
which is the type of the underlying elements.