Add client metrics (#751)
* WIP: add client metrics

* remove toolchain

* reflect revivew comments

* fix golangci

* remove golint

* address review comments

* Apply suggestions from code review

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>

* [MM-58179] Review networking performance at scale (#752)

* Enable ethtool metrics

* Increase network interface rx size on proxy instance

* Use retransmission rate instead of timeouts

* Review socket buffer sizes

* Update panel unit

* Add TCP retransmissions panel

* Fix datasource

* Make RX size dynamic

* Fix comment

* Update postgres client (#755)

* Update postgres client

* Make assets

* Bump version strings in master (#757)

* Update retransmission threshold (#760)

Half the value based on latest test results

* Update coverage frequency docs (#761)

Recover broken images and tweak things that have changed since this was
first written.

* MM-59319: Allow opensearch installations to be created (#763)

* MM-59319: Allow opensearch installations to be created

We unblock the prefix limitation, and allow
either prefixes.

https://mattermost.atlassian.net/browse/MM-59319

* Fix test

* Allow 0 shard replicas (#764)

* Allow 0 shard replicas

For 1-node ES clusters, we need 0 shard replicas, not 1. Otherwise, the
cluster status check is never valid, since there are always unassigned
shards (one per index) and subsequently the cluster status is always
yellow.

* Validate RestoreSnapshot options before marshaling

* MM-59316: Remove cloudwatch log policy (#762)

* Remove Cloudwatch log policy

There's a limit of 10 Cloudwatch log policies per region per account, so
it's not scalable to create one with every deployment. Instead, we rely
on such a policy already present in the AWS account. If it is not
present, the only downside is that logs cannot be viewable through
Cloudwatch, but everything else should keep working.

* make assets

* Check needed policy and create if it doesn't exist

* Refactor CloudWatch logic to another file and test

* Improvements from Redis investigation (#765)

* Improvements from Redis investigation

- Turn off the cache_lock. This is unnecessary and
we shouldn't even have done this to start with.
- Set timeout for entire response and not just response header.
It is unclear why exactly this causes the problem. So open to
thoughts and suggestions.

* bumped lingering timeout

* feat: allow saml with the external auth provider (#759)

* allow saml authentication using keycloak

* validate saml authentication type

* handle saml redirect

* Update loadtest/user/userentity/actions.go

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>

* Update api/agent.go

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>

* close bodies

* gofmt

* better error message

---------

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>

* Revert the disabling of proxy_cache_lock (#768)

Release testing for v1.19 of the load-test tool showed that disabling
proxy_cache_lock was affecting the performance of the app nodes, which
were eventually killed by the overhead of the requests that were no
longer waiting for the cache to be updated.

This commit reverts that change, and a subsequent investigation into the
proxy_cache_lock_timeout fine-tuning will follow.
See https://mattermost.atlassian.net/browse/MM-59871

* Bump v1.19.0 rc2 (#770)

* Bump version to v1.19.0-rc1 (#766)

* Bump version to v1.19.0-rc2

* Run make update-dependencies (#772)

* Fix race condition in RoundTrip test (#774)

There was a race condition where the test ended up creating the SSH
client *before* the SSH server was up.

Using require.Eventually, we check every 100ms, up to 5s, that the
client is correctly created, giving time to the server to start.

* Bump version to v1.19.0 (#775)

* Add local file option for Mattermost download url (#767)

* allow binaries with specific mattermost builds; don't fail on license check

* lint

* nit

* full stop on license error (instead of outputting error and continuing)

* allow three kinds of MattermostDownloadURL

* Fix incorrect deployer.sample.json (#776)

The newly added Elasticsearch config settings don't accept string values.

* Update the load test dashboard panels to new format (#777)

* Update the load test dashboard panels to new format

* Address review comments

* reflect reviews

* fix the TTFB measurement

* remove ttb observation from controller actions

* revert TTFB measurement to be calculated only on login

* remove random platform/agent

---------

Co-authored-by: Alejandro García Montoro <alejandro.garciamontoro@gmail.com>
Co-authored-by: Claudio Costa <cstcld91@gmail.com>
Co-authored-by: Agniva De Sarker <agnivade@yahoo.co.in>
Co-authored-by: Felipe Martin <812088+fmartingr@users.noreply.github.com>
Co-authored-by: Christopher Poile <cpoile@gmail.com>