Ignore error from aws s3 rm command (#864)
This command is making the whole {comparison, deployment} destroy to
fail when there is an S3 bucket deployed. That error is valid in the
local context of the aws s3 rm command, but not in the global sense:
since we run the aws s3 rm command concurrently with terraform destroy,
aws s3 rm fails when the bucket is destroyed before the command has
finished.

This is a spurious failure: all resources, including the S3 bucket, have
been destroyed (that's exactly why the aws s3 rm command fails,
actually), so we can simply skip it.

In the case this command would error in any other way that we are
not accounting for as of today, the {comparison, deployment} destroy
command would still fail, since it would notice that the S3 bucket has
not been destroyed.