Remove OpenTelemetry collector's debug exporter
This was useful for debugging the collector when setting it up, but it's
no longer needed, and if there's any unrelated error (like an agent
being down), /var/log/syslog can quickly end up consuming the whole
disk. The error logs will still be sent to Loki, so it's not like we're
removing useful info for debugging potential errors.