Use -version:creatordate to pick latest tag
When there is more than one tag in a commit, git sorts them
lexicographically by default, so v1.26.0-rc1 is listed *after* v1.26.0.
As goreleaser picks the latest one listed by git, we need to change the
sorting method, so that v1.26.0 is listed after v1.26.0-rc1. We do this
by setting tag_sort to -version:creatordate, so that they are sorted by
creation date. As the stable tag will always be created after the
release candidate tag, this will work as expected.

More info in the docs: https://goreleaser.com/customization/git/