MM-63542: Add load-test coverage for new RHS loading pattern (#904)
Our existing viewThread implementation was incorrect. It used
to load pages of posts in a thread as a user scrolls up or down.

That is the ideal way, but our webapp actually loads the entire
thread. That is the first change done here.

Then, we apply the performance optimization where for subsequent
loads of a thread, we use the lastUpdateAt of a thread and pass
that to the API call.

https://mattermost.atlassian.net/browse/MM-63542