MM-63724: Remove excessive calls to GetChannelsForTeam (#908)
* Remove excessive calls to GetChannelsForTeam

The webapp calls GetChannelsForTeam for a single team during the initial
load phase, which is already done by loadTeam.

* Add GetAllChannelMembersForUser

* Add call to GetAllChannelMembersForUser to Reload

* Remove extra calls to get channels and memberships

This is removed from the webapp in this PR:
https://github.com/mattermost/mattermost/pull/30771/files#diff-3cbe2755372a6058ac41ad749710ec3f63cd72fb6ff3cafec22cdca472fc3b40L33

* Call methods regardless of having a current team

Both GetTeamMembersForUser and GetChannelsForUser do not depend on
having set a current team, so we can move them out of the teamId != ""
check.

* Return semver.Version instead of string from Store

We should parse the server version string as soon as possible, so that
we store an already parsed semver.Version and not a simple string,
removing the need for callers to parse it over and over again.

* Version GetAllChannelMembersForUser

Starting with v10.9.0, the GetChannelMembersWithTeamData client function
supports streaming all members directly, without the need for
pagination, by setting the page param to -1. This commit adds that
possibility, keeping the pagination logic for older servers

* Use perPage instead of hard-coded number