MM-63743: Add flag to maintain metrics instance (#905)
* Allow Terraform.Destroy to forget about resources

This lets the caller specify a list of resources that should be removed
from Terraform state, thus avoiding they get destroyed.

* Add --do-not-destroy-metrics-instance to destroy

Also add a util function to use as PreRunE to make sure that the user
undrestand what they are doing.

* Add flag to comparison destroy command as well

* Add EnableMetricsInstance setting

This allows us to decouple the creation of metrics-related AWS resources
from the creation of app instances, so that we can remove the latter and
still keep the former.

* Set non-metrics resource counts to 0 to destroy

Instead of removing the metrics-related resources from Terraform state,
set all other instance counts to 0 and run create again, effectively
destroying those resources.

* Revert "Allow Terraform.Destroy to forget about resources"

This reverts commit 083d7161e77458e2c94e51a56e2749180d52693e.

* Fix spurious error when there are no DB instances

* Setup agents only if there are agents

This effectively allows a deployment with no agents.

* Reword warning message

* Use the same logic in the comparison command

* Refactor config destroy logic to util method

* Override comparison cfg in deploymentAction

There is no need to run another loop before that, we can use
the *deploymentConfig received by the passed function.

* make assets

* Revert "Override comparison cfg in deploymentAction"

This reverts commit ff8fa2f0520c010c76a5ce035435bc7bbc5b2e94.

* Reword message

* Refactor askForConfirmation function

* Test askForConfirmation function

* Rename DestroyAllButMetrics