Use private IPs in the metrics' hosts file
The security group rules that reference the ID of another security
groups are designed to work only with private IPs, not public ones.
Thus, for those rules to apply, the metrics instance needs to contact
all resources via their private IPs, regardless of the ConnectionType
setting set by the user.