Revert change from #908
This reverts the change done in PR #908 that made the GetChannelsForUser
call to be executed unconditionally. This prevented new users with no
team membership to finish the Reload function. We need a better solution
here, but we're going back to the old behaviour to unblock the
performance comparison for v10.9-rc1.