Refresh the AWS credentials if a role is provided
When an AWS role is provided, the credentials chain is most probably
using the role chaining logic, which allows for a maximum duration of
one hour. Comparisons usually need more than one hour to finish, and
they make use of the AWS SDK after that time. This logic allows for the
credentials to be refreshed automatically, assuming the role over and
over again while it is needed.

See https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles.html#iam-term-role-chaining